requirements.txt
streamlit==1.31.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
supabase==2.3.4
pandas==2.2.0
plotly==5.18.0
python-telegram-bot==20.7
python-dotenv==1.0.1
toml==0.10.2
